requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
requests>=2.31.0
opencv-python>=4.8.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
spotipy>=2.23.0
streamlit>=1.28.0
